maj7: - frequency params now have a KT center of 0.3, and default to this value. now corresponds better with massive, and more comfortable range, and saves a mul in osc. - added ability to clear sample in vst editor - master vol default to -6db which is just more natural - clamp01 and N11 are now functions due to frequency of use. - fixed embarrassing typo making sqrt(0) = 0.5.

+++ WaveSabreCore/src/Helpers.cpp
@@ -127,7 +127,7 @@ namespace WaveSabreCore
 
 	double Helpers::Exp2(double x)
 	{
-		return M7::math::pow2(x);
+		return (double)M7::math::pow2((float)x);
 	}
 	float Helpers::Exp2F(float x)
 	{
